+++ Modules/DFSDsc/DSCResources/MSFT_DFSReplicationGroup/MSFT_DFSReplicationGroup.psm1
@@ -626,13 +626,18 @@ function Test-TargetResource
 
             # Compare the Members
             $existingMembers = @((Get-DfsrMember @replicationGroupParameters -ErrorAction Stop).DnsName)
-            #check for null values before using compare-object
-            # if one is null but not both then report difference
-            # if neither are null then compare values
-            if ([String]::IsNullOrEmpty($existingMembers) -and [String]::IsNullOrEmpty($fqdnMembers)) {
-                #both are null so no difference found
-            } elseif ([String]::IsNullOrEmpty($existingMembers) -xor [String]::IsNullOrEmpty($fqdnMembers)) {
-                # There is a member different of some kind.
+            <#
+                check for null values before using compare-object
+                if one is null but not both then report difference
+                if neither are null then compare values
+            #>
+            if ([String]::IsNullOrEmpty($existingMembers) -and [String]::IsNullOrEmpty($fqdnMembers))
+            {
+                # both are null so no difference found
+            }
+            elseif ([String]::IsNullOrEmpty($existingMembers) -xor [String]::IsNullOrEmpty($fqdnMembers))
+            {
+                # There is a member difference of some kind.
                 Write-Verbose -Message ( @(
                     "$($MyInvocation.MyCommand): "
                     $($LocalizedData.ReplicationGroupMembersNeedUpdateMessage) `
@@ -640,13 +645,15 @@ function Test-TargetResource
                     ) -join '' )
 
                 $desiredConfigurationMatch = $false
-            }  else {
-                #neither variables are null so it's safe to use compare-object
+            }
+            else
+            {
+                # neither variables are null so it's safe to use compare-object
                 if ((Compare-Object `
                     -ReferenceObject $fqdnMembers `
                     -DifferenceObject $existingMembers).Count -ne 0)
                 {
-                    # There is a member different of some kind.
+                    # There is a member difference of some kind.
                     Write-Verbose -Message ( @(
                         "$($MyInvocation.MyCommand): "
                         $($LocalizedData.ReplicationGroupMembersNeedUpdateMessage) `
@@ -659,13 +666,18 @@ function Test-TargetResource
 
             # Compare the Folders
             $existingFolders = @((Get-DfsReplicatedFolder @replicationGroupParameters -ErrorAction Stop).FolderName)
-            #check for null values before using compare-object
-            # if one is null but not both then report difference
-            # if neither are null then compare values
-            if ([String]::IsNullOrEmpty($existingFolders) -and [String]::IsNullOrEmpty($Folders)){
-                #both are null so no difference found
-            } elseif ([String]::IsNullOrEmpty($existingFolders) -xor [String]::IsNullOrEmpty($Folders)) {
-                # There is a folder different of some kind.
+            <#
+                check for null values before using compare-object
+                if one is null but not both then report difference
+                if neither are null then compare values
+            #>
+            if ([String]::IsNullOrEmpty($existingFolders) -and [String]::IsNullOrEmpty($Folders))
+            {
+                # both are null so no difference found
+            }
+            elseif ([String]::IsNullOrEmpty($existingFolders) -xor [String]::IsNullOrEmpty($Folders))
+            {
+                # There is a folder difference of some kind.
                 Write-Verbose -Message ( @(
                     "$($MyInvocation.MyCommand): "
                     $($LocalizedData.ReplicationGroupFoldersNeedUpdateMessage) `
@@ -673,13 +685,15 @@ function Test-TargetResource
                     ) -join '' )
 
                 $desiredConfigurationMatch = $false
-            } else {
-                #neither variables are null so it's safe to use compare-object
+            }
+            else
+            {
+                # neither variables are null so it's safe to use compare-object
                 if ((Compare-Object `
                     -ReferenceObject $Folders `
                     -DifferenceObject $existingFolders).Count -ne 0)
                 {
-                    # There is a folder different of some kind.
+                    # There is a folder difference of some kind.
                     Write-Verbose -Message ( @(
                         "$($MyInvocation.MyCommand): "
                         $($LocalizedData.ReplicationGroupFoldersNeedUpdateMessage) `